#277 - extracting inner class TooManyStars.Submitter
